fix: buffer overflows in slave mailbox handlers, div-by-zero in autotuning
Fix 15 issues identified in the second-phase audit of the remaining
Tether framework modules (control, slave stack, G-code, simulation,
destabilizer, identification, drives, profiles):

Slave mailbox buffer overflows (HIGH):
- FoEHandler sendDataPacket assumed 512-byte buffer but caller provides
  256 — use responseLen input as buffer capacity
- EoEHandler handleGetIPParam: add bounds check for DNS name copy
- SoEHandler handleReadRequest: add bounds checks for all variable-length
  field memcpy operations
- AoEHandler handleRead: add total response size check before writing
- CoEHandler processGetObjDesc/EntryDesc: clamp name length to buffer
  capacity in addition to protocol limit
- SlaveCore: add safety clamp on responseLen after handler returns

CiA401 I/O buffer safety (HIGH):
- processDigitalInputs: replace hardcoded 16 with actual len parameter,
  fix unsigned underflow when offset > 16
- processAnalogInputs: replace hardcoded 32 with actual len parameter

Control autotuning div-by-zero (HIGH):
- PIDController::autoTune: validate processGain and deadTime non-zero
  before dividing
- AstromHagglundRelay: validate amplitude non-zero before computing Ku

Documentation and interface safety (MEDIUM/LOW):
- VariableSystem: correct false claim of mutex protection
- CompositeControllers: document raw pointer lifetime and thread safety
- ForwardKinematics Quaternion::normalized: guard against zero magnitude
- ESCConfig: widen processDataRamSize return type to uint32_t
- CohenCoon/LambdaTuning/AMIGOMethod: restore conditional error messages

+++ include/tether/slave/core/ESCConfig.hpp
@@ -31,7 +31,7 @@ struct ESCConfig {
     EtherCAT::ESC::ESCFeatureReg features = std::bit_cast<EtherCAT::ESC::ESCFeatureReg>(
         static_cast<uint16_t>(ESCFeature::FMMU | ESCFeature::SyncManager | ESCFeature::DC));
 
-    uint16_t processDataRamSize() const { return ramSizeKB * 1024; }
+    uint32_t processDataRamSize() const { return static_cast<uint32_t>(ramSizeKB) * 1024; }
 };
 
 }} // namespace EtherCAT::slave
